requirements.txt
langchain
langchain-community
langchain-openai
openai
chromadb
tiktoken
python-dotenv
streamlit
